rework tests, remove mocha
