+++ example/index.ts
@@ -2,9 +2,6 @@
 import { install } from 'react-native-quick-crypto';
 install();
 
-// mocha things
-global.process.cwd = () => 'sxsx';
-
 // readable-stream
 // @ts-expect-error - although process.version is readonly, we're setting it for readable-stream
 global.process.version = 'v22.0.0';
