+++ example/package.json
@@ -24,9 +24,9 @@
     "@react-navigation/native": "6.1.18",
     "@react-navigation/native-stack": "6.11.0",
     "buffer": "6.0.3",
+    "chai": "<5.0.0",
     "crypto-browserify": "^3.12.0",
     "events": "3.3.0",
-    "mocha": "10.7.0",
     "react": "18.3.1",
     "react-native": "0.76.1",
     "react-native-bouncy-checkbox": "4.0.1",
@@ -54,7 +54,8 @@
     "@react-native/eslint-plugin": "0.76.1",
     "@react-native/metro-config": "0.76.1",
     "@react-native/typescript-config": "0.76.1",
-    "@types/mocha": "10.0.9",
+    "@tsconfig/react-native": "^3.0.5",
+    "@types/chai": "^5.0.1",
     "@types/react": "18.3.3",
     "@types/react-native-vector-icons": "^6.4.18",
     "@types/react-test-renderer": "18.3.0",
@@ -65,7 +66,6 @@
     "eslint-plugin-react-native": "^4.1.0",
     "jest": "29.7.0",
     "prettier": "3.3.3",
-    "react-test-renderer": "18.3.1",
     "typescript": "5.0.4",
     "typescript-eslint": "^8.0.1"
   },
