+++ example/src/hooks/useTestsList.ts
@@ -1,26 +1,26 @@
 import { useState, useCallback } from 'react';
 import type { Suites, TestSuite } from '../types/Suite';
-import { rootSuite } from '../testing/MochaRNAdapter';
+import { TestsContext } from '../tests/util';
 
-import '../testing/tests/pbkdf2/pbkdf2_tests';
-import '../testing/tests/random/random_tests';
-// import '../testing/tests/HmacTests/HmacTests';
-// import '../testing/tests/HashTests/HashTests';
-// import '../testing/tests/CipherTests/CipherTestFirst';
-// import '../testing/tests/CipherTests/CipherTestSecond';
-// import '../testing/tests/CipherTests/PublicCipherTests';
-// import '../testing/tests/CipherTests/test398';
-// import '../testing/tests/CipherTests/generateKey';
-// import '../testing/tests/CipherTests/GenerateKeyPairTests';
-// import '../testing/tests/ConstantsTests/ConstantsTests';
-// import '../testing/tests/SignTests/SignTests';
-// import '../testing/tests/SmokeTests/bundlerTests';
-// import '../testing/tests/webcryptoTests/deriveBits';
-// import '../testing/tests/webcryptoTests/digest';
-// import '../testing/tests/webcryptoTests/generateKey';
-// import '../testing/tests/webcryptoTests/encrypt_decrypt';
-// import '../testing/tests/webcryptoTests/import_export';
-// import '../testing/tests/webcryptoTests/sign_verify';
+import '../tests/pbkdf2/pbkdf2_tests';
+import '../tests/random/random_tests';
+// import '../tests/HmacTests/HmacTests';
+// import '../tests/HashTests/HashTests';
+// import '../tests/CipherTests/CipherTestFirst';
+// import '../tests/CipherTests/CipherTestSecond';
+// import '../tests/CipherTests/PublicCipherTests';
+// import '../tests/CipherTests/test398';
+// import '../tests/CipherTests/generateKey';
+// import '../tests/CipherTests/GenerateKeyPairTests';
+// import '../tests/ConstantsTests/ConstantsTests';
+// import '../tests/SignTests/SignTests';
+// import '../tests/SmokeTests/bundlerTests';
+// import '../tests/webcryptoTests/deriveBits';
+// import '../tests/webcryptoTests/digest';
+// import '../tests/webcryptoTests/generateKey';
+// import '../tests/webcryptoTests/encrypt_decrypt';
+// import '../tests/webcryptoTests/import_export';
+// import '../tests/webcryptoTests/sign_verify';
 
 export const useTestsList = (): [
   Suites<TestSuite>,
@@ -28,13 +28,13 @@ export const useTestsList = (): [
   () => void,
   () => void,
 ] => {
-  const [suites, setSuites] = useState<Suites<TestSuite>>(getInitialSuites());
+  const [suites, setSuites] = useState<Suites<TestSuite>>(TestsContext);
 
   const toggle = useCallback(
     (description: string) => {
-      setSuites(tests => {
-        tests[description]!.value = !tests[description]!.value;
-        return tests;
+      setSuites(suites => {
+        suites[description]!.value = !suites[description]!.value;
+        return suites;
       });
     },
     [setSuites],
@@ -59,17 +59,4 @@ export const useTestsList = (): [
   }, [setSuites]);
 
   return [suites, toggle, clearAll, checkAll];
-};
-
-const getInitialSuites = () => {
-  const suites: Suites = {};
-
-  // interrogate the loaded mocha suites/tests via a temporary runner
-  const runner = new Mocha.Runner(rootSuite);
-  runner.suite.suites.map(s => {
-    suites[s.title] = { value: false, count: s.total() };
-  });
-
-  // return count-enhanced list and totals
-  return suites;
 };
